backend/requirements.txt
fastapi
uvicorn
openai
azure-cognitiveservices-speech
motor
python-dotenv
